nuevas implementaciones en los post

+++ src/main/java/com/example/testbackenddeveloper/services/impl/CandidateServiceImp.java
@@ -47,7 +47,9 @@ public class CandidateServiceImp implements CandidateService {
 
     @Override
     public void deleteCandidate(Long id) {
-        candidateRepository.deleteById(id);
+        if (this.findById(id) != null) {
+            candidateRepository.deleteById(id);
+        }
     }
 
     @Override
